split pkg
