+++ frontend/configs/vitest/index.ts
@@ -1,6 +1,6 @@
 /// <reference types="vitest" />
-import { defineConfig } from "vitest/config";
 import { vanillaExtractPlugin } from "@vanilla-extract/vite-plugin";
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   plugins: [vanillaExtractPlugin()],
