+++ frontend/foundation/testing/browser/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@foundation/testing",
+  "name": "@foundation/testing/browser",
   "version": "0.0.0",
   "main": "./dist/src/index.js",
   "module": "./dist/src/index.js",
@@ -17,20 +17,15 @@
     "clean": "../../scripts/clean.sh"
   },
   "dependencies": {
-    "@foundation/rpc": "workspace:*",
-    "@codegen/rpc": "workspace:*",
-    "@configs/vitest": "workspace:*"
+    "@foundation/testing/rpc": "workspace:*"
   },
   "peerDependencies": {
-    "@tanstack/query-core": "*",
-    "@tanstack/react-query": "*",
-    "msw": "*",
-    "react": "*",
-    "vitest": "*"
+    "msw": "*"
   },
   "devDependencies": {
     "@configs/eslint": "workspace:*",
-    "@configs/typescript": "workspace:*"
+    "@configs/typescript": "workspace:*",
+    "@configs/vitest": "workspace:*"
   },
   "type": "module"
 }
