+++ frontend/foundation/testing/browser/src/index.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from "msw/browser";
-import { mockRPCHandlers, RPCMocks } from "./msw.shared";
+import { mockRPCHandlers, RPCMocks } from "@foundation/testing/rpc";
 
 export const mockRPCsWorker = (mocks: RPCMocks): void => {
   const worker = setupWorker(...mockRPCHandlers(mocks));
