+++ frontend/foundation/testing/rpc/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./msw";
 export * from "./PageTestWrap";
 export * from "./rpc";
