+++ frontend/foundation/testing/server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@product/login",
+  "name": "@foundation/testing/server",
   "version": "0.0.0",
   "main": "./dist/src/index.js",
   "module": "./dist/src/index.js",
@@ -17,21 +17,10 @@
     "clean": "../../scripts/clean.sh"
   },
   "dependencies": {
-    "@foundation/ui": "workspace:*",
-    "@codegen/rpc": "workspace:*",
-    "@foundation/testing/rpc": "workspace:*",
-    "@foundation/testing/server": "workspace:*",
-    "@foundation/rpc": "workspace:*",
-    "@foundation/icons": "workspace:*",
-    "@foundation/forms": "workspace:*",
-    "@foundation/layout": "workspace:*",
-    "@foundation/routing": "workspace:*"
+    "@foundation/testing/rpc": "workspace:*"
   },
   "peerDependencies": {
-    "@testing-library/react": "*",
-    "@testing-library/user-event": "*",
-    "jotai": "*",
-    "react": "*",
+    "msw": "*",
     "vitest": "*"
   },
   "devDependencies": {
