+++ frontend/foundation/testing/server/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "@configs/typescript/base.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": [
+    "src",
+    ".eslintrc.cjs",
+    "vitest.config.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist"
+  ]
+}
