+++ frontend/foundation/testing/server/vitest.config.ts
@@ -1,0 +1,1 @@
+export { default } from "@configs/vitest";
