+++ frontend/product/login/src/page/index.test.tsx
@@ -1,5 +1,5 @@
-import { PageTestWrap } from "@foundation/testing";
-import { mockRPCsForTest } from "@foundation/testing/dist/src/msw.server";
+import { PageTestWrap } from "@foundation/testing/rpc";
+import { mockRPCsForTest } from "@foundation/testing/server";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it, vi } from "vitest";
