+++ frontend/stories/package.json
@@ -15,7 +15,8 @@
   "dependencies": {
     "@foundation/theme": "workspace:*",
     "@foundation/ui": "workspace:*",
-    "@foundation/testing": "workspace:*",
+    "@foundation/testing/rpc": "workspace:*",
+    "@foundation/testing/browser": "workspace:*",
     "@foundation/layout": "workspace:*",
     "@foundation/loaders": "workspace:*",
     "@foundation/icons": "workspace:*",
